+++ app/pb_kits/playbook/pb_textarea/docs/index.js
@@ -1,4 +1,5 @@
 export { default as TextareaDefault } from './_textarea_default.jsx'
+export { default as TextareaResize } from './_textarea_resize.jsx'
 export { default as TextareaCustom } from './_textarea_custom.jsx'
 export { default as TextareaDark } from './_textarea_dark.jsx'
 export { default as TextareaError } from './_textarea_error.jsx'